refactor(obj-dept): clarify lifecycle operations
- plans/obj-dept-remediation.md: Update remediation guidance for clear/delete operations and constructors.
- plans/obj-dept-review-issues.md: Record resolved lifecycle, navigation, cleanup, and performance findings.
- poprako-obj-dept-macro/src/impl_obj_dept/lifecycle.rs: Generate separate batch cleanup steps and optimize ownership handling.
- poprako-obj-dept/src/lib.rs: Remove obj_inst! and register ClearObjs/DeleteObjs capabilities.
- poprako-obj-dept/src/model.rs: Remove the obsolete upload-mark result module.
- poprako-obj-dept/src/model/mark.rs: Delete the replaced upload-mark outcome enum.
- poprako-obj-dept/src/oper.rs: Add constructors, private markers, and split cleanup operations.
- poprako-obj-dept/src/tests.rs: Update operation construction and lifecycle expectations.
- rust-style-lint.toml: Extend style checks for revised operation APIs.
- scripts/ci-check.sh: Adjust CI checking for the updated source layout.
- specs/obj-dept.md: Specify constructor-based operations and business-oriented cleanup semantics.
- src/benchmark.rs: Replace macro-built operations with associated constructors.
- src/complex/user.rs: Add revised object-department operation imports.
- src/part_impl/obj_dept/mock_impl.rs: Implement renamed cleanup operations and optimize replacement handling.
- src/part_impl/obj_dept/mock_impl/tests.rs: Update mock lifecycle tests for split cleanup operations.
- src/part_impl/prom/mock_impl/chapter.rs: Adapt chapter promotion cleanup to the new operation names.
- src/usecase/chapter/delete.rs: Use DeleteObjs for chapter image cleanup.
- src/usecase/chapter/stage.rs: Use constructors and ClearObjs for chapter staging.
- src/usecase/chapter_port/export.rs: Use direct object operation constructors during export.
- src/usecase/comic.rs: Update comic object operation construction and imports.
- src/usecase/comic/cover.rs: Use direct cover object operation constructors.
- src/usecase/comic/reserve.rs: Update comic reservation object operations.
- src/usecase/comic_archive.rs: Clear active covers and delete archived page objects explicitly.
- src/usecase/internal/view.rs: Use direct metadata and URL operation constructors.
- src/usecase/page.rs: Handle boolean upload marking at the use-case boundary.
- src/usecase/page/delete.rs: Use DeleteObjs for page deletion cleanup.
- src/usecase/page/reserve.rs: Replace macro operations and batch-delete removed page objects.
- src/usecase/page/view.rs: Borrow model IDs during metadata and URL lookup.
- src/usecase/prom.rs: Use direct metadata operation construction in promotion checks.
- src/usecase/team.rs: Handle team avatar marking and reservation with direct constructors.
- src/usecase/team/delete.rs: Use DeleteObjs for team avatar cleanup.
- src/usecase/team/view.rs: Use direct avatar metadata and URL constructors.
- src/usecase/user.rs: Handle user avatar marking and reservation with direct constructors.
- src/usecase/user/delete.rs: Use DeleteObjs for user avatar cleanup.
- src/usecase/user/view.rs: Use direct avatar metadata and URL constructors.

+++ src/benchmark.rs
@@ -26,10 +26,10 @@ use crate::value::role::{RoleField, RoleMask};
 const CHAPTER_COUNT: usize = 8;
 
 // Number of pages included in each benchmark chapter snapshot.
-const PAGE_COUNT: usize = 8;
+const PAGE_COUNT: u16 = 8;
 
 // Number of units included on each benchmark page.
-const UNIT_COUNT: usize = 48;
+const UNIT_COUNT: u16 = 48;
 
 /// Benchmarks archive preparation for a large comic snapshot.
 pub struct ArchiveInput(ComicArchiveSnapshot);
@@ -182,13 +182,13 @@ fn archive_snapshot() -> Option<ComicArchiveSnapshot> {
             updated_at: archived_at,
         };
 
-        let mut page_snapshots = Vec::with_capacity(PAGE_COUNT);
+        let mut page_snapshots = Vec::with_capacity(usize::from(PAGE_COUNT));
 
         for page_index in 0..PAGE_COUNT {
             //
             let page_id = format!("page-{}-{}", chapter_index, page_index);
 
-            let mut unit_infos = Vec::with_capacity(UNIT_COUNT);
+            let mut unit_infos = Vec::with_capacity(usize::from(UNIT_COUNT));
 
             for unit_index in 0..UNIT_COUNT {
                 //
@@ -205,10 +205,10 @@ fn archive_snapshot() -> Option<ComicArchiveSnapshot> {
                 page_info: PageInfo {
                     id: page_id,
                     chapter_id: chapter_id.clone(),
-                    index: page_index,
-                    total_unit_count: UNIT_COUNT,
-                    translated_unit_count: UNIT_COUNT,
-                    proofread_unit_count: UNIT_COUNT,
+                    index: usize::from(page_index),
+                    total_unit_count: usize::from(UNIT_COUNT),
+                    translated_unit_count: usize::from(UNIT_COUNT),
+                    proofread_unit_count: usize::from(UNIT_COUNT),
                     created_at: archived_at,
                     updated_at: archived_at,
                 },
@@ -224,10 +224,10 @@ fn archive_snapshot() -> Option<ComicArchiveSnapshot> {
                 is_pinned: chapter_index == 0,
                 index: chapter_index,
                 subtitle: format!("Chapter {}", chapter_index),
-                page_count: PAGE_COUNT,
-                total_unit_count: PAGE_COUNT * UNIT_COUNT,
-                translated_unit_count: PAGE_COUNT * UNIT_COUNT,
-                proofread_unit_count: PAGE_COUNT * UNIT_COUNT,
+                page_count: usize::from(PAGE_COUNT),
+                total_unit_count: usize::from(PAGE_COUNT * UNIT_COUNT),
+                translated_unit_count: usize::from(PAGE_COUNT * UNIT_COUNT),
+                proofread_unit_count: usize::from(PAGE_COUNT * UNIT_COUNT),
                 stages,
                 creator_id: "user-1".into(),
                 creator: None,
@@ -286,7 +286,7 @@ fn export_input() -> LabelPlusExportInput {
     //
     let archived_at = OffsetDateTime::UNIX_EPOCH;
 
-    let mut pages = Vec::with_capacity(PAGE_COUNT * 8);
+    let mut pages = Vec::with_capacity(usize::from(PAGE_COUNT) * 8);
 
     let mut units_by_page_id = HashMap::new();
 
@@ -303,10 +303,10 @@ fn export_input() -> LabelPlusExportInput {
         pages.push(PageInfo {
             id: page_id.clone(),
             chapter_id: "chapter-1".into(),
-            index: page_index,
-            total_unit_count: UNIT_COUNT,
-            translated_unit_count: UNIT_COUNT,
-            proofread_unit_count: UNIT_COUNT,
+            index: usize::from(page_index),
+            total_unit_count: usize::from(UNIT_COUNT),
+            translated_unit_count: usize::from(UNIT_COUNT),
+            proofread_unit_count: usize::from(UNIT_COUNT),
             created_at: archived_at,
             updated_at: archived_at,
         });
@@ -338,8 +338,8 @@ fn user_info(archived_at: OffsetDateTime) -> UserInfo {
 fn unit_info(
     page_id: &str,
     chapter_index: usize,
-    page_index: usize,
-    unit_index: usize,
+    page_index: u16,
+    unit_index: u16,
     archived_at: OffsetDateTime,
 ) -> UnitInfo {
     //
@@ -352,8 +352,8 @@ fn unit_info(
         is_bubble: unit_index.is_multiple_of(2),
         is_proofread: true,
         coord: UnitCoord {
-            x_coord: unit_index as f64,
-            y_coord: page_index as f64,
+            x_coord: f64::from(unit_index),
+            y_coord: f64::from(page_index),
         },
         translated_text: Some(format!(
             "Translated text for chapter {}, page {}, unit {}.",
